changed long id types to int

+++ app/src/main/java/com/doozez/doozez/api/invitation/InvitationCreateRequest.kt
@@ -4,8 +4,8 @@ import com.google.gson.annotations.SerializedName
 
 class InvitationCreateRequest {
     @SerializedName("safe")
-    var safeId: Long? = null
+    var safeId: Int? = null
 
     @SerializedName("recipient")
-    var recipientId: Long? = null
+    var recipientId: Int? = null
 }
